Move fuel cell tech unlocks to Power nodes
That's where the actual fuel cell parts are. Fuel cells are often
used for life support, but aren't inherently related to life support.

As a side-effect, this means a Gemini capsule still has to research
the fuel cell Power node to be able to use its built-in fuel cells,
putting it on equal footing with a capsule without built-in cells
(like the RO Advanced Capsule)