Fix Prototype Spaceplane habitat size
Apply the patch :after[RO], because the parts get created in :for[RO]
(I didn't think it mattered for +patches, but without this the patch
never gets applied)

Also use %, in case the 1.25m part gets patched before the +part[]
clones it. (not sure this is needed, was the 1st thing I tried and it
wasn't enough. can't hurt)