Update ROKerbalism.cfg
Changed Algae to use CO2.  It's not strong enough to replace a scrubber, and should run all the time instead of competing with the real scrubber.

Scaled MRP to produce enough Oxygen for 3 crew.  Fuel production sized process can be added to dedicated ISRU part when we get there.

Removed Waste Incinerator to cut down on the process clutter in the chemical plant.  The new Waste Processor fills the same role.