Autodump water, oxygen
Set fuel cells to always dump water, and KO2 scrubbers to always dump oxygen.

Forgetting to set resource dumping is a major trap for new players, and this avoids any possible bugs with kerbalism forgetting to dump (if those haven't been fixed yet).

Although there are potentially niche scenarios in which the player would not want this to happen, breaking those are preferable to the fuel cell and KO2 scrubber not functioning as expected 99% of the time.